int/double weight

+++ lib/models/dto/request/create_pet_req_dto.dart
@@ -12,7 +12,7 @@ class CreatePetReqDto {
   late String? dateOfBirth; // ISO String;
   late String? colour;
   late String? notes;
-  late int? weight;
+  late double? weight;
   late String? avatar;
 
   CreatePetReqDto({
